Small update to fix issue #40
This answer on Gurobi Google Groups (https://groups.google.com/forum/#!topic/gurobi/8JiUTt7CcMg) suggests that Gurobi 'Slack' variables are signed based on the 'Sense' attribute of the constraint. This may have led to a small sign error in getconstrsolution which I attempted to fix. I hope this works :)