Update grb_model.jl
There is a problem in the tune_model wrapper. A call to the gurobi function "tunemodel" does not affect the model but only computes an array of parameters settings. A subsequent call to "gettuneresult" is necessary if we want to apply the best parameters found to the given model.

Another way to accomplish this would be to encapsulate the call to "gettuneresult" within another julia wrapper.

Best wishes