Update README.md
This is an attempt to make the README file consistent.  At the beginning of the linear programming section, it says 

In all examples below, we assume that the preamble codes like the following exist in the script:

```julia
using Gurobi
env = Gurobi.Env()
... optional codes to set parameters to env ...
```

However, neither Example 1.3 nor 1.4 make use of the modified environment.