Adds section Common Performance Pitfall with JuMP. (#301)
As discussed with @odow in https://github.com/JuliaOpt/Gurobi.jl/pull/300, this updates the documentation of README.md to give new JuMP.jl+Gurobi.jl users a warning about the most common performance pitfall (so they may be already aware of it and structure their JuMP programs accordingly). This will also probably help users that find the "excessive updates" message and search for how to fix it.